pyproject.toml
[project]
name = "argus-overview"
version = "2.7.0"
description = "Professional multi-boxing tool for EVE Online on Linux"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AreteDriver"}
]
keywords = ["eve-online", "multiboxing", "linux", "gaming", "preview", "pyside6", "argus", "overview"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
]
requires-python = ">=3.9"
dependencies = [
    "PySide6>=6.6.0",
    "python-xlib>=0.33",
    "Pillow>=10.0.0",
    "pynput>=1.7.0",
    "numpy>=1.24.0",
    "watchdog>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "isort",
    "pytest",
    "pytest-cov",
    "bandit[toml]",
]

[project.urls]
Homepage = "https://github.com/AreteDriver/Argus_Overview"
Repository = "https://github.com/AreteDriver/Argus_Overview.git"
Issues = "https://github.com/AreteDriver/Argus_Overview/issues"
Changelog = "https://github.com/AreteDriver/Argus_Overview/blob/main/CHANGELOG.md"

[project.scripts]
argus-overview = "main:main"

[tool.ruff]
line-length = 100
target-version = "py38"
exclude = ["windows/", "build/", "dist/", ".venv/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "N802",  # Function name should be lowercase (Qt method overrides use camelCase)
    "N806",  # Variable in function should be lowercase (Mock class variables)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F811", "B017"]  # Duplicate class definitions, bare exceptions in test assertions

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.bandit]
exclude_dirs = ["tests", "windows", ".venv", "build", "dist", "benchmarks"]
skips = [
    "B101",  # assert used (tests)
    "B404",  # subprocess import (required for X11 tools)
    "B603",  # subprocess without shell=True (this is the safe way)
    "B607",  # partial path (system tools: wmctrl, xdotool, xrandr)
]
